refactor(map component): cleaned up marker rendering logic
