--- frontend/src/datas/data.jsx
@@ -1,2111 +1,0 @@
-export const mapData = {
-    "Classroom Block 1": {
-        "type": "FeatureCollection",
-        "features": [
-          {
-            "type": "Feature",
-            "properties": {},
-            "geometry": {
-              "coordinates": [
-                76.92481299254285,
-                10.927741509257018
-              ],
-              "type": "Point"
-            }
-          }
-        ]
-    },
-    "ES Block": {
-        "type": "FeatureCollection",
-        "features": [
-          {
-            "type": "Feature",
-            "properties": {
-              "name": "ES Block",
-              "description": "Electrical Sciences Block."
-            },
-            "geometry": {
-              "coordinates": [
-                76.92415320802343,
-                10.927585988849472
-              ],
-              "type": "Point"
-            }
-          }
-        ]
-    },
-    "Main Block": {
-        "type": "FeatureCollection",
-        "features": [
-          {
-            "type": "Feature",
-            "properties": {},
-            "geometry": {
-              "coordinates": [
-                76.92578800992112,
-                10.927670081305351
-              ],
-              "type": "Point"
-            }
-          }
-        ]
-    },
-    "CSE/IT Block": {
-        "type": "FeatureCollection",
-        "features": [
-          {
-            "type": "Feature",
-            "properties": {},
-            "geometry": {
-              "coordinates": [
-                76.92529732115082,
-                10.926189663956094
-              ],
-              "type": "Point"
-            }
-          }
-        ]
-    },
-    "Aroma": {
-        "type": "FeatureCollection",
-        "features": [
-          {
-            "type": "Feature",
-            "properties": {
-              "name": "Aroma",
-              "description": "It is food court in the campus. With various amenities around it.",
-              "landmark": "Backside to SKASC Placement Block."
-            },
-            "geometry": {
-              "coordinates": [
-                76.92322666294586,
-                10.926982511920599
-              ],
-              "type": "Point"
-            }
-          }
-        ]
-    },
-    "Library": {
-      "type": "FeatureCollection",
-      "features": [
-        {
-          "type": "Feature",
-          "properties": {},
-          "geometry": {
-            "coordinates": [
-              76.92490230768863,
-              10.928395378639863
-            ],
-            "type": "Point"
-          }
-        }
-      ]
-  },
-  "MBA Block": {
-    "type": "FeatureCollection",
-    "features": [
-      {
-        "type": "Feature",
-        "properties": {},
-        "geometry": {
-          "coordinates": [
-            76.9242531396186,
-            10.928550420705776
-          ],
-          "type": "Point"
-        }
-      }
-    ]
-  },
-  "Parking": {
-    "type": "FeatureCollection",
-    "features": [
-      {
-        "type": "Feature",
-        "properties": {},
-        "geometry": {
-          "coordinates": [
-            76.92301089990355,
-            10.928587577014142
-          ],
-          "type": "Point"
-        }
-      }
-    ]
-  },
-  "Vankatram Hall": {
-    "type": "FeatureCollection",
-    "features": [
-      {
-        "type": "Feature",
-        "properties": {},
-        "geometry": {
-          "coordinates": [
-            76.92550466804448,
-            10.925504934718234
-          ],
-          "type": "Point"
-        }
-      }
-    ]
-  },
-  "Classroom Block 2": {
-    "type": "FeatureCollection",
-    "features": [
-      {
-        "type": "Feature",
-        "properties": {},
-        "geometry": {
-          "coordinates": [
-            76.92526622009473,
-            10.927104791045693
-          ],
-          "type": "Point"
-        }
-      }
-    ]
-  },
-  "EEE Block": {
-    "type": "FeatureCollection",
-    "features": [
-      {
-        "type": "Feature",
-        "properties": {},
-        "geometry": {
-          "coordinates": [
-            76.92463035699677,
-            10.927182832461796
-          ],
-          "type": "Point"
-        }
-      }
-    ]
-  }
-}
-
-
-export const mapLayout = {
-  "MBA Block": {
-    "type": "FeatureCollection",
-    "features": [
-      {
-        "type": "Feature",
-        "properties": {},
-        "geometry": {
-          "coordinates": [
-            [
-              76.92389293084256,
-              10.928681600273578
-            ],
-            [
-              76.9239168846616,
-              10.92845111014367
-            ],
-            [
-              76.92398874612206,
-              10.92845581402598
-            ],
-            [
-              76.92398899649737,
-              10.92842834476464
-            ],
-            [
-              76.92403894079757,
-              10.928430573789342
-            ],
-            [
-              76.92403667060324,
-              10.928403825488417
-            ],
-            [
-              76.92443023697257,
-              10.928433274108741
-            ],
-            [
-              76.92443818805816,
-              10.92840204654307
-            ],
-            [
-              76.92445250001299,
-              10.928408292056929
-            ],
-            [
-              76.92447158261945,
-              10.92828494314223
-            ],
-            [
-              76.9246526931729,
-              10.928306403197439
-            ],
-            [
-              76.92464653561643,
-              10.928375930935587
-            ],
-            [
-              76.92459727515853,
-              10.928366862101015
-            ],
-            [
-              76.92459419637908,
-              10.928430343936157
-            ],
-            [
-              76.92465577195122,
-              10.92843638982498
-            ],
-            [
-              76.9246496143947,
-              10.928521032250146
-            ],
-            [
-              76.92464037805877,
-              10.92874472997093
-            ],
-            [
-              76.92452646324926,
-              10.928741707030227
-            ],
-            [
-              76.92451722691334,
-              10.928811234666355
-            ],
-            [
-              76.92444025744874,
-              10.928814257606348
-            ],
-            [
-              76.9244371786694,
-              10.928920060499024
-            ],
-            [
-              76.92402770111346,
-              10.928889831105124
-            ],
-            [
-              76.92403385866993,
-              10.928793097023544
-            ],
-            [
-              76.92397536187605,
-              10.928793097023544
-            ],
-            [
-              76.92397536187605,
-              10.92869636291043
-            ],
-            [
-              76.92388607729612,
-              10.92867822526054
-            ]
-          ],
-          "type": "LineString"
-        }
-      }
-    ]
-  },
-  // "ES Block":
-  // {
-  //   "type": "FeatureCollection",
-  //   "features": [
-  //     {
-  //       "type": "Feature",
-  //       "properties": {},
-  //       "geometry": {
-  //         "coordinates": [
-  //           [
-  //             76.92414424651543,
-  //             10.927788650409198
-  //           ],
-  //           [
-  //             76.92414424651543,
-  //             10.9277063077664
-  //           ],
-  //           [
-  //             76.92401598494382,
-  //             10.9277063077664
-  //           ],
-  //           [
-  //             76.92402091808242,
-  //             10.927541622413116
-  //           ],
-  //           [
-  //             76.92404558376904,
-  //             10.927362405894883
-  //           ],
-  //           [
-  //             76.92401105180744,
-  //             10.927362405894883
-  //           ],
-  //           [
-  //             76.92404558376904,
-  //             10.927202564043867
-  //           ],
-  //           [
-  //             76.92445010102983,
-  //             10.927236469898503
-  //           ],
-  //           [
-  //             76.9244303684797,
-  //             10.927347874821493
-  //           ],
-  //           [
-  //             76.92454383063773,
-  //             10.927367249586197
-  //           ],
-  //           [
-  //             76.92450929867812,
-  //             10.92773536987825
-  //           ],
-  //           [
-  //             76.92441063592958,
-  //             10.927720838823177
-  //           ],
-  //           [
-  //             76.92426264181,
-  //             10.927711151452044
-  //           ],
-  //           [
-  //             76.92427250808498,
-  //             10.927803181461016
-  //           ],
-  //           [
-  //             76.92413438024039,
-  //             10.927793494093535
-  //           ]
-  //         ],
-  //         "type": "LineString"
-  //       }
-  //     }
-  //   ]
-  // },
-// "Classroom Block 1":{
-//   "type": "FeatureCollection",
-//   "features": [
-//     {
-//       "type": "Feature",
-//       "properties": {},
-//       "geometry": {
-//         "coordinates": [
-//           [
-//             76.92401765783393,
-//             10.927697697050888
-//           ],
-//           [
-//             76.9240485751784,
-//             10.92736610799912
-//           ],
-//           [
-//             76.92403668389153,
-//             10.92736610799912
-//           ],
-//           [
-//             76.9240513883941,
-//             10.927202640035873
-//           ],
-//           [
-//             76.92423927072593,
-//             10.927218985992823
-//           ],
-//           [
-//             76.92439742483975,
-//             10.927235331948012
-//           ],
-//           [
-//             76.92442012426676,
-//             10.926993854053862
-//           ],
-//           [
-//             76.92482111424181,
-//             10.927040517956371
-//           ],
-//           [
-//             76.9248077478914,
-//             10.927098117774193
-//           ],
-//           [
-//             76.92480712895104,
-//             10.927164359204525
-//           ],
-//           [
-//             76.92457316904841,
-//             10.92715159709536
-//           ],
-//           [
-//             76.92458121529012,
-//             10.927164966924934
-//           ],
-//           [
-//             76.9246158760144,
-//             10.927168005521821
-//           ],
-//           [
-//             76.9246158760144,
-//             10.927197176055074
-//           ],
-//           [
-//             76.92466043980585,
-//             10.927259163428104
-//           ],
-//           [
-//             76.92479504538318,
-//             10.927264283036706
-//           ],
-//           [
-//             76.92477276318289,
-//             10.927363948927422
-//           ],
-//           [
-//             76.92477276318289,
-//             10.927416212761997
-//           ],
-//           [
-//             76.92476471681522,
-//             10.927422289820157
-//           ],
-//           [
-//             76.92463969062533,
-//             10.927409527722077
-//           ],
-//           [
-//             76.9245487062193,
-//             10.927402842813393
-//           ],
-//           [
-//             76.92453760637017,
-//             10.927510841726345
-//           ],
-//           [
-//             76.92453017907093,
-//             10.927603214964222
-//           ],
-//           [
-//             76.92453021435227,
-//             10.927611688328156
-//           ],
-//           [
-//             76.92451969234492,
-//             10.927612903765066
-//           ],
-//           [
-//             76.92451659763873,
-//             10.927738701480038
-//           ],
-//           [
-//             76.92430182484793,
-//             10.927728370038906
-//           ],
-//           [
-//             76.92412418855321,
-//             10.927716215584908
-//           ],
-//           [
-//             76.92412392951599,
-//             10.92770927558864
-//           ],
-//           [
-//             76.92401725085676,
-//             10.927707723820646
-//           ],
-//           [
-//             76.92401725085676,
-//             10.927699189102299
-//           ]
-//         ],
-//         "type": "LineString"
-//       }
-//     },
-//     {
-//       "type": "Feature",
-//       "properties": {},
-//       "geometry": {
-//         "coordinates": [
-//           [
-//             76.92470018268097,
-//             10.92773150294839
-//           ],
-//           [
-//             76.92460439033016,
-//             10.927724971326484
-//           ],
-//           [
-//             76.92461503392508,
-//             10.927621771680933
-//           ],
-//           [
-//             76.9247081653765,
-//             10.927630915955419
-//           ],
-//           [
-//             76.9247161480722,
-//             10.927542085854526
-//           ],
-//           [
-//             76.9248598365977,
-//             10.927551230130334
-//           ],
-//           [
-//             76.9248558452494,
-//             10.927582581932796
-//           ],
-//           [
-//             76.92487048019268,
-//             10.92759041988262
-//           ],
-//           [
-//             76.9248558452494,
-//             10.927602176807554
-//           ],
-//           [
-//             76.92488378468568,
-//             10.927611321082594
-//           ],
-//           [
-//             76.92488511513483,
-//             10.927556455431173
-//           ],
-//           [
-//             76.92499155107936,
-//             10.927562987056788
-//           ],
-//           [
-//             76.92498090748444,
-//             10.927647898176815
-//           ],
-//           [
-//             76.92507936073338,
-//             10.927658348774955
-//           ],
-//           [
-//             76.92509266522643,
-//             10.927655736125999
-//           ],
-//           [
-//             76.92507803028428,
-//             10.927760242083778
-//           ],
-//           [
-//             76.92496627254229,
-//             10.927751097813271
-//           ],
-//           [
-//             76.9249649420932,
-//             10.927836008879467
-//           ],
-//           [
-//             76.9249649420932,
-//             10.927847765794638
-//           ],
-//           [
-//             76.92468953908605,
-//             10.92782555828866
-//           ],
-//           [
-//             76.92469885223187,
-//             10.927732809272541
-//           ]
-//         ],
-//         "type": "LineString"
-//       }
-//     }
-//   ]
-// },
-// "Main Block":{
-//   "type": "FeatureCollection",
-//   "features": [
-//     {
-//       "type": "Feature",
-//       "properties": {},
-//       "geometry": {
-//         "coordinates": [
-//           [
-//             76.92470018268097,
-//             10.92773150294839
-//           ],
-//           [
-//             76.92460439033016,
-//             10.927724971326484
-//           ],
-//           [
-//             76.92461503392508,
-//             10.927621771680933
-//           ],
-//           [
-//             76.9247081653765,
-//             10.927630915955419
-//           ],
-//           [
-//             76.9247161480722,
-//             10.927542085854526
-//           ],
-//           [
-//             76.9248598365977,
-//             10.927551230130334
-//           ],
-//           [
-//             76.9248558452494,
-//             10.927582581932796
-//           ],
-//           [
-//             76.92487048019268,
-//             10.92759041988262
-//           ],
-//           [
-//             76.9248558452494,
-//             10.927602176807554
-//           ],
-//           [
-//             76.92488378468568,
-//             10.927611321082594
-//           ],
-//           [
-//             76.92488511513483,
-//             10.927556455431173
-//           ],
-//           [
-//             76.92499155107936,
-//             10.927562987056788
-//           ],
-//           [
-//             76.92498090748444,
-//             10.927647898176815
-//           ],
-//           [
-//             76.92507936073338,
-//             10.927658348774955
-//           ],
-//           [
-//             76.92509266522643,
-//             10.927655736125999
-//           ],
-//           [
-//             76.92507803028428,
-//             10.927760242083778
-//           ],
-//           [
-//             76.92496627254229,
-//             10.927751097813271
-//           ],
-//           [
-//             76.9249649420932,
-//             10.927836008879467
-//           ],
-//           [
-//             76.9249649420932,
-//             10.927847765794638
-//           ],
-//           [
-//             76.92468953908605,
-//             10.92782555828866
-//           ],
-//           [
-//             76.92469885223187,
-//             10.927732809272541
-//           ]
-//         ],
-//         "type": "LineString"
-//       }
-//     },
-//     {
-//       "type": "Feature",
-//       "properties": {},
-//       "geometry": {
-//         "coordinates": [
-//           [
-//             76.92402106068147,
-//             10.9276942780549
-//           ],
-//           [
-//             76.92405526643552,
-//             10.927367582487847
-//           ],
-//           [
-//             76.92402840997465,
-//             10.927359671635912
-//           ],
-//           [
-//             76.92404452384909,
-//             10.927196180659223
-//           ],
-//           [
-//             76.92415021313917,
-//             10.927202354940718
-//           ],
-//           [
-//             76.92415033542977,
-//             10.927181566731008
-//           ],
-//           [
-//             76.92432565992124,
-//             10.927196868567293
-//           ],
-//           [
-//             76.92433734822032,
-//             10.927227472238314
-//           ],
-//           [
-//             76.92444838706587,
-//             10.927237035884687
-//           ],
-//           [
-//             76.92443669876508,
-//             10.927353712345592
-//           ],
-//           [
-//             76.92455163370983,
-//             10.927363275987872
-//           ],
-//           [
-//             76.92452436101087,
-//             10.927652097841332
-//           ],
-//           [
-//             76.92451462076133,
-//             10.927743908702482
-//           ],
-//           [
-//             76.92426916647366,
-//             10.927732432346644
-//           ],
-//           [
-//             76.92426527037452,
-//             10.927803203202416
-//           ],
-//           [
-//             76.92413864713069,
-//             10.927793639574276
-//           ],
-//           [
-//             76.92414059518023,
-//             10.927715217811354
-//           ],
-//           [
-//             76.924015919986,
-//             10.9277056541804
-//           ],
-//           [
-//             76.9240178680356,
-//             10.927682701464576
-//           ]
-//         ],
-//         "type": "LineString"
-//       }
-//     },
-//     {
-//       "type": "Feature",
-//       "properties": {},
-//       "geometry": {
-//         "coordinates": [
-//           [
-//             76.92574552181708,
-//             10.927712073439181
-//           ],
-//           [
-//             76.92575361247776,
-//             10.927626675908613
-//           ],
-//           [
-//             76.92530255806736,
-//             10.927588942107775
-//           ],
-//           [
-//             76.92529851273707,
-//             10.927662423714963
-//           ],
-//           [
-//             76.92519737946128,
-//             10.927656465747702
-//           ],
-//           [
-//             76.92520951545407,
-//             10.927436020867091
-//           ],
-//           [
-//             76.9257657484722,
-//             10.927493614590645
-//           ],
-//           [
-//             76.92576777113737,
-//             10.927443964829905
-//           ],
-//           [
-//             76.92595790169605,
-//             10.927459852754097
-//           ],
-//           [
-//             76.92595183369889,
-//             10.927509502512152
-//           ],
-//           [
-//             76.92650806671702,
-//             10.927553194292557
-//           ],
-//           [
-//             76.92648986272707,
-//             10.927751793212536
-//           ],
-//           [
-//             76.92638670678446,
-//             10.92774583524701
-//           ],
-//           [
-//             76.92639682011202,
-//             10.927668381682153
-//           ],
-//           [
-//             76.9259396977061,
-//             10.927642563822985
-//           ],
-//           [
-//             76.92592958437854,
-//             10.927729947338165
-//           ],
-//           [
-//             76.92574147648503,
-//             10.927710087450706
-//           ]
-//         ],
-//         "type": "LineString"
-//       }
-//     }
-//   ]
-// },
-"CSE/IT Block":{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92470018268097,
-            10.92773150294839
-          ],
-          [
-            76.92460439033016,
-            10.927724971326484
-          ],
-          [
-            76.92461503392508,
-            10.927621771680933
-          ],
-          [
-            76.9247081653765,
-            10.927630915955419
-          ],
-          [
-            76.9247161480722,
-            10.927542085854526
-          ],
-          [
-            76.9248598365977,
-            10.927551230130334
-          ],
-          [
-            76.9248558452494,
-            10.927582581932796
-          ],
-          [
-            76.92487048019268,
-            10.92759041988262
-          ],
-          [
-            76.9248558452494,
-            10.927602176807554
-          ],
-          [
-            76.92488378468568,
-            10.927611321082594
-          ],
-          [
-            76.92488511513483,
-            10.927556455431173
-          ],
-          [
-            76.92499155107936,
-            10.927562987056788
-          ],
-          [
-            76.92498090748444,
-            10.927647898176815
-          ],
-          [
-            76.92507936073338,
-            10.927658348774955
-          ],
-          [
-            76.92509266522643,
-            10.927655736125999
-          ],
-          [
-            76.92507803028428,
-            10.927760242083778
-          ],
-          [
-            76.92496627254229,
-            10.927751097813271
-          ],
-          [
-            76.9249649420932,
-            10.927836008879467
-          ],
-          [
-            76.9249649420932,
-            10.927847765794638
-          ],
-          [
-            76.92468953908605,
-            10.92782555828866
-          ],
-          [
-            76.92469885223187,
-            10.927732809272541
-          ]
-        ],
-        "type": "LineString"
-      }
-    },
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92402106068147,
-            10.9276942780549
-          ],
-          [
-            76.92405526643552,
-            10.927367582487847
-          ],
-          [
-            76.92402840997465,
-            10.927359671635912
-          ],
-          [
-            76.92404452384909,
-            10.927196180659223
-          ],
-          [
-            76.92415021313917,
-            10.927202354940718
-          ],
-          [
-            76.92415033542977,
-            10.927181566731008
-          ],
-          [
-            76.92432565992124,
-            10.927196868567293
-          ],
-          [
-            76.92433734822032,
-            10.927227472238314
-          ],
-          [
-            76.92444838706587,
-            10.927237035884687
-          ],
-          [
-            76.92443669876508,
-            10.927353712345592
-          ],
-          [
-            76.92455163370983,
-            10.927363275987872
-          ],
-          [
-            76.92452436101087,
-            10.927652097841332
-          ],
-          [
-            76.92451462076133,
-            10.927743908702482
-          ],
-          [
-            76.92426916647366,
-            10.927732432346644
-          ],
-          [
-            76.92426527037452,
-            10.927803203202416
-          ],
-          [
-            76.92413864713069,
-            10.927793639574276
-          ],
-          [
-            76.92414059518023,
-            10.927715217811354
-          ],
-          [
-            76.924015919986,
-            10.9277056541804
-          ],
-          [
-            76.9240178680356,
-            10.927682701464576
-          ]
-        ],
-        "type": "LineString"
-      }
-    },
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92574552181708,
-            10.927712073439181
-          ],
-          [
-            76.92575361247776,
-            10.927626675908613
-          ],
-          [
-            76.92530255806736,
-            10.927588942107775
-          ],
-          [
-            76.92529851273707,
-            10.927662423714963
-          ],
-          [
-            76.92519737946128,
-            10.927656465747702
-          ],
-          [
-            76.92520951545407,
-            10.927436020867091
-          ],
-          [
-            76.9257657484722,
-            10.927493614590645
-          ],
-          [
-            76.92576777113737,
-            10.927443964829905
-          ],
-          [
-            76.92595790169605,
-            10.927459852754097
-          ],
-          [
-            76.92595183369889,
-            10.927509502512152
-          ],
-          [
-            76.92650806671702,
-            10.927553194292557
-          ],
-          [
-            76.92648986272707,
-            10.927751793212536
-          ],
-          [
-            76.92638670678446,
-            10.92774583524701
-          ],
-          [
-            76.92639682011202,
-            10.927668381682153
-          ],
-          [
-            76.9259396977061,
-            10.927642563822985
-          ],
-          [
-            76.92592958437854,
-            10.927729947338165
-          ],
-          [
-            76.92574147648503,
-            10.927710087450706
-          ]
-        ],
-        "type": "LineString"
-      }
-    },
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92523087549819,
-            10.926315869720938
-          ],
-          [
-            76.92523555813574,
-            10.926042303563975
-          ],
-          [
-            76.92531710382286,
-            10.92604323457904
-          ],
-          [
-            76.9253161556174,
-            10.925948270810494
-          ],
-          [
-            76.92545649005234,
-            10.925951063862513
-          ],
-          [
-            76.92545649005234,
-            10.925941753687496
-          ],
-          [
-            76.92547545416505,
-            10.925944546740354
-          ],
-          [
-            76.92547355775412,
-            10.925952925897889
-          ],
-          [
-            76.9254934700723,
-            10.925951063862513
-          ],
-          [
-            76.92547450595964,
-            10.925899857896567
-          ],
-          [
-            76.92552855368154,
-            10.925848651921825
-          ],
-          [
-            76.925618633218,
-            10.925952925897889
-          ],
-          [
-            76.92560915116002,
-            10.925980856420864
-          ],
-          [
-            76.92558734243079,
-            10.925982718455288
-          ],
-          [
-            76.92558923884178,
-            10.926016235079075
-          ],
-          [
-            76.92560630654356,
-            10.926017166096187
-          ],
-          [
-            76.92559208345983,
-            10.926234093042183
-          ],
-          [
-            76.92550484827098,
-            10.926234092922655
-          ],
-          [
-            76.92550200365292,
-            10.926319746436647
-          ],
-          [
-            76.92541097591254,
-            10.926319746436647
-          ],
-          [
-            76.9252289204286,
-            10.92632067745356
-          ]
-        ],
-        "type": "LineString"
-      }
-    }
-  ]
-},
-"Aroma":{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92332385543824,
-            10.927083787129334
-          ],
-          [
-            76.92345377182329,
-            10.927094879381656
-          ],
-          [
-            76.92349896013047,
-            10.926832824820664
-          ],
-          [
-            76.92339587430445,
-            10.926820346026062
-          ],
-          [
-            76.92336763161217,
-            10.926824505624694
-          ],
-          [
-            76.92338316509262,
-            10.926683079253024
-          ],
-          [
-            76.92322359388095,
-            10.92666505431923
-          ],
-          [
-            76.92314451434288,
-            10.92665534858412
-          ],
-          [
-            76.92314592647722,
-            10.926641483248702
-          ],
-          [
-            76.92308520468839,
-            10.926635937114156
-          ],
-          [
-            76.9230739076122,
-            10.92663732364747
-          ],
-          [
-            76.92307249547662,
-            10.92665812165066
-          ],
-          [
-            76.92307108334222,
-            10.926678919653554
-          ],
-          [
-            76.92309932603456,
-            10.926677533119232
-          ],
-          [
-            76.92309226536145,
-            10.926757952046287
-          ],
-          [
-            76.92318687838122,
-            10.926770430843519
-          ],
-          [
-            76.92318405411123,
-            10.92680370763317
-          ],
-          [
-            76.92315016288035,
-            10.926802321099473
-          ],
-          [
-            76.92304990132305,
-            10.926785682705585
-          ],
-          [
-            76.92303154357376,
-            10.926990889504296
-          ],
-          [
-            76.9231727570351,
-            10.927000595227241
-          ],
-          [
-            76.92322076961216,
-            10.927006141354994
-          ],
-          [
-            76.92327866713106,
-            10.927014460545905
-          ],
-          [
-            76.9232885520729,
-            10.926986729907966
-          ],
-          [
-            76.92330267341907,
-            10.926979797248094
-          ],
-          [
-            76.92331679476519,
-            10.926958999267455
-          ],
-          [
-            76.92330973409207,
-            10.92701030095104
-          ],
-          [
-            76.92332385543824,
-            10.927015847077499
-          ],
-          [
-            76.92331961903511,
-            10.927082400598067
-          ]
-        ],
-        "type": "LineString"
-      }
-    }
-  ]
-},
-"Library":{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92479601628247,
-            10.92848126684271
-          ],
-          [
-            76.92484596261932,
-            10.928483847923218
-          ],
-          [
-            76.92484990575048,
-            10.928516111433893
-          ],
-          [
-            76.92497214283736,
-            10.928529016837189
-          ],
-          [
-            76.92497082846035,
-            10.928519983055352
-          ],
-          [
-            76.92506940675588,
-            10.928529016837189
-          ],
-          [
-            76.92507072113295,
-            10.928467070895877
-          ],
-          [
-            76.92506155116774,
-            10.928467889410783
-          ],
-          [
-            76.92506738660137,
-            10.928357389854483
-          ],
-          [
-            76.9250650280818,
-            10.928308180289307
-          ],
-          [
-            76.92501629305838,
-            10.928306730253624
-          ],
-          [
-            76.92495869711985,
-            10.928340081104935
-          ],
-          [
-            76.92490405481982,
-            10.928341531141612
-          ],
-          [
-            76.92489962436252,
-            10.928398082575711
-          ],
-          [
-            76.9248361211479,
-            10.928398082575711
-          ],
-          [
-            76.92483907478652,
-            10.928338631067092
-          ],
-          [
-            76.92483907478652,
-            10.928321230624022
-          ],
-          [
-            76.92475046565062,
-            10.928315430475692
-          ],
-          [
-            76.92474751201189,
-            10.928385032245544
-          ],
-          [
-            76.92479477021857,
-            10.928389382356045
-          ],
-          [
-            76.92478886294253,
-            10.928480734651487
-          ],
-          [
-            76.9248080615892,
-            10.928482184687425
-          ]
-        ],
-        "type": "LineString"
-      }
-    }
-  ]
-},
-"Parking":{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92260339101989,
-            10.929084246995473
-          ],
-          [
-            76.9232875912486,
-            10.929011838417779
-          ],
-          [
-            76.92345147154282,
-            10.928951497922597
-          ],
-          [
-            76.92358667278694,
-            10.928790589875476
-          ],
-          [
-            76.92361535183704,
-            10.928577386578112
-          ],
-          [
-            76.92352112066823,
-            10.928585431987827
-          ],
-          [
-            76.92283692043952,
-            10.928685999597903
-          ],
-          [
-            76.92280824138766,
-            10.928637727149408
-          ],
-          [
-            76.92260339101989,
-            10.928669908782055
-          ],
-          [
-            76.92261568161729,
-            10.928702090313266
-          ],
-          [
-            76.922066682631,
-            10.92879058977671
-          ],
-          [
-            76.92220188387512,
-            10.929007815619528
-          ],
-          [
-            76.9222920180373,
-            10.929039997212044
-          ],
-          [
-            76.9222920180373,
-            10.92909631499181
-          ],
-          [
-            76.92264436066921,
-            10.92908022419904
-          ]
-        ],
-        "type": "LineString"
-      }
-    }
-  ]
-},
-"Vankatram Hall":{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92260339101989,
-            10.929084246995473
-          ],
-          [
-            76.9232875912486,
-            10.929011838417779
-          ],
-          [
-            76.92345147154282,
-            10.928951497922597
-          ],
-          [
-            76.92358667278694,
-            10.928790589875476
-          ],
-          [
-            76.92361535183704,
-            10.928577386578112
-          ],
-          [
-            76.92352112066823,
-            10.928585431987827
-          ],
-          [
-            76.92283692043952,
-            10.928685999597903
-          ],
-          [
-            76.92280824138766,
-            10.928637727149408
-          ],
-          [
-            76.92260339101989,
-            10.928669908782055
-          ],
-          [
-            76.92261568161729,
-            10.928702090313266
-          ],
-          [
-            76.922066682631,
-            10.92879058977671
-          ],
-          [
-            76.92220188387512,
-            10.929007815619528
-          ],
-          [
-            76.9222920180373,
-            10.929039997212044
-          ],
-          [
-            76.9222920180373,
-            10.92909631499181
-          ],
-          [
-            76.92264436066921,
-            10.92908022419904
-          ]
-        ],
-        "type": "LineString"
-      }
-    },
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92525270319499,
-            10.925631798726187
-          ],
-          [
-            76.92526499421672,
-            10.92531400169446
-          ],
-          [
-            76.92537168712471,
-            10.925310485194842
-          ],
-          [
-            76.92537168712471,
-            10.925238075696413
-          ],
-          [
-            76.92571173873489,
-            10.925242098446958
-          ],
-          [
-            76.9257076417282,
-            10.925310485194842
-          ],
-          [
-            76.92585103698565,
-            10.925322553443436
-          ],
-          [
-            76.92583055194888,
-            10.925672532424514
-          ],
-          [
-            76.92544133625006,
-            10.925652418701148
-          ],
-          [
-            76.92544133625006,
-            10.925696668890865
-          ],
-          [
-            76.92534710507954,
-            10.925696668890865
-          ],
-          [
-            76.92535939610298,
-            10.92563632772081
-          ],
-          [
-            76.92526106792576,
-            10.925632304976446
-          ]
-        ],
-        "type": "LineString"
-      }
-    }
-  ]
-},
-"Classroom Block 2":{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92260339101989,
-            10.929084246995473
-          ],
-          [
-            76.9232875912486,
-            10.929011838417779
-          ],
-          [
-            76.92345147154282,
-            10.928951497922597
-          ],
-          [
-            76.92358667278694,
-            10.928790589875476
-          ],
-          [
-            76.92361535183704,
-            10.928577386578112
-          ],
-          [
-            76.92352112066823,
-            10.928585431987827
-          ],
-          [
-            76.92283692043952,
-            10.928685999597903
-          ],
-          [
-            76.92280824138766,
-            10.928637727149408
-          ],
-          [
-            76.92260339101989,
-            10.928669908782055
-          ],
-          [
-            76.92261568161729,
-            10.928702090313266
-          ],
-          [
-            76.922066682631,
-            10.92879058977671
-          ],
-          [
-            76.92220188387512,
-            10.929007815619528
-          ],
-          [
-            76.9222920180373,
-            10.929039997212044
-          ],
-          [
-            76.9222920180373,
-            10.92909631499181
-          ],
-          [
-            76.92264436066921,
-            10.92908022419904
-          ]
-        ],
-        "type": "LineString"
-      }
-    },
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92525270319499,
-            10.925631798726187
-          ],
-          [
-            76.92526499421672,
-            10.92531400169446
-          ],
-          [
-            76.92537168712471,
-            10.925310485194842
-          ],
-          [
-            76.92537168712471,
-            10.925238075696413
-          ],
-          [
-            76.92571173873489,
-            10.925242098446958
-          ],
-          [
-            76.9257076417282,
-            10.925310485194842
-          ],
-          [
-            76.92585103698565,
-            10.925322553443436
-          ],
-          [
-            76.92583055194888,
-            10.925672532424514
-          ],
-          [
-            76.92544133625006,
-            10.925652418701148
-          ],
-          [
-            76.92544133625006,
-            10.925696668890865
-          ],
-          [
-            76.92534710507954,
-            10.925696668890865
-          ],
-          [
-            76.92535939610298,
-            10.92563632772081
-          ],
-          [
-            76.92526106792576,
-            10.925632304976446
-          ]
-        ],
-        "type": "LineString"
-      }
-    },
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "coordinates": [
-          [
-            76.92504207670993,
-            10.927149654227094
-          ],
-          [
-            76.92495328023864,
-            10.927120592057804
-          ],
-          [
-            76.92496807965017,
-            10.92703340553291
-          ],
-          [
-            76.92505194298508,
-            10.92702856183719
-          ],
-          [
-            76.92506180926006,
-            10.926946218983502
-          ],
-          [
-            76.92513087318133,
-            10.926941375285267
-          ],
-          [
-            76.92513087318133,
-            10.92690746939796
-          ],
-          [
-            76.92509634122177,
-            10.926888094602262
-          ],
-          [
-            76.92510620749476,
-            10.926800908010122
-          ],
-          [
-            76.92528873357605,
-            10.926810595409933
-          ],
-          [
-            76.92527393416452,
-            10.926912313095642
-          ],
-          [
-            76.92522460279133,
-            10.92690746939796
-          ],
-          [
-            76.92522460279133,
-            10.926946218983502
-          ],
-          [
-            76.92534793122434,
-            10.92696559377434
-          ],
-          [
-            76.9253331318128,
-            10.927057624014452
-          ],
-          [
-            76.92543672769574,
-            10.927072155103033
-          ],
-          [
-            76.92542686142275,
-            10.927169029004688
-          ],
-          [
-            76.92531833240128,
-            10.927169029004688
-          ],
-          [
-            76.92531339926268,
-            10.927256215489734
-          ],
-          [
-            76.92503714357343,
-            10.927231997024307
-          ],
-          [
-            76.92505194298508,
-            10.927139966837274
-          ]
-        ],
-        "type": "LineString"
-      }
-    }
-  ]
-},
-// "EEE Block":{
-//   "type": "FeatureCollection",
-//   "features": [
-//     {
-//       "type": "Feature",
-//       "properties": {},
-//       "geometry": {
-//         "coordinates": [
-//           [
-//             76.92414424651543,
-//             10.927788650409198
-//           ],
-//           [
-//             76.92414424651543,
-//             10.9277063077664
-//           ],
-//           [
-//             76.92401598494382,
-//             10.9277063077664
-//           ],
-//           [
-//             76.92402091808242,
-//             10.927541622413116
-//           ],
-//           [
-//             76.92404558376904,
-//             10.927362405894883
-//           ],
-//           [
-//             76.92401105180744,
-//             10.927362405894883
-//           ],
-//           [
-//             76.92404558376904,
-//             10.927202564043867
-//           ],
-//           [
-//             76.92445010102983,
-//             10.927236469898503
-//           ],
-//           [
-//             76.9244303684797,
-//             10.927347874821493
-//           ],
-//           [
-//             76.92454383063773,
-//             10.927367249586197
-//           ],
-//           [
-//             76.92450929867812,
-//             10.92773536987825
-//           ],
-//           [
-//             76.92441063592958,
-//             10.927720838823177
-//           ],
-//           [
-//             76.92426264181,
-//             10.927711151452044
-//           ],
-//           [
-//             76.92427250808498,
-//             10.927803181461016
-//           ],
-//           [
-//             76.92413438024039,
-//             10.927793494093535
-//           ]
-//         ],
-//         "type": "LineString"
-//       }
-//     },
-//     {
-//       "type": "Feature",
-//       "properties": {},
-//       "geometry": {
-//         "coordinates": [
-//           [
-//             76.92455113483595,
-//             10.92740176505032
-//           ],
-//           [
-//             76.92476930222114,
-//             10.927417065879126
-//           ],
-//           [
-//             76.92478488560573,
-//             10.927270177890193
-//           ],
-//           [
-//             76.92464463514472,
-//             10.92725487705384
-//           ],
-//           [
-//             76.92455736819005,
-//             10.927150831344932
-//           ],
-//           [
-//             76.92479423563623,
-//             10.927160011849878
-//           ],
-//           [
-//             76.92480981902071,
-//             10.927046785600908
-//           ],
-//           [
-//             76.92440776769848,
-//             10.926997822884033
-//           ],
-//           [
-//             76.92438906763624,
-//             10.927224275378691
-//           ]
-//         ],
-//         "type": "LineString"
-//       }
-//     }
-//   ]
-// }
-  }
-
-
-
-export const minimumPath = {
-  "Classroom Block 1": {
-    "type": "FeatureCollection",
-    "features": [
-      {
-        "type": "Feature",
-        "properties": {},
-        "geometry": {
-          "coordinates": [
-            [
-              76.9252057220969,
-              10.926064290568462
-            ],
-            [
-              76.92516397696073,
-              10.926209613056074
-            ],
-            [
-              76.9248489909416,
-              10.926429459761508
-            ],
-            [
-              76.92492109617467,
-              10.926623222823565
-            ],
-            [
-              76.92483298675586,
-              10.927526326970536
-            ]
-          ],
-          "type": "LineString"
-        }
-      }
-    ]
-  },
-  "Aroma": {
-    "type": "FeatureCollection",
-    "features": [
-      {
-        "type": "Feature",
-        "properties": {},
-        "geometry": {
-          "coordinates": [
-            [
-              76.92511344582738,
-              10.926215832597293
-            ],
-            [
-              76.92504504044513,
-              10.926214799284637
-            ],
-            [
-              76.92503030697742,
-              10.926196199652708
-            ],
-            [
-              76.92502925458626,
-              10.926114567918972
-            ],
-            [
-              76.92474226919325,
-              10.92610821541517
-            ],
-            [
-              76.9243777914881,
-              10.92610821541517
-            ],
-            [
-              76.92430481865762,
-              10.926275565688925
-            ],
-            [
-              76.92424696649874,
-              10.926606361827695
-            ],
-            [
-              76.92390165371245,
-              10.927245502592186
-            ],
-            [
-              76.92321798843574,
-              10.927158796932758
-            ],
-            [
-              76.92323223146309,
-              10.92701055816562
-            ]
-          ],
-          "type": "LineString"
-        }
-      }
-    ]
-  }
-}
