+++ frontend/vite.config.js
@@ -6,7 +6,7 @@ export default defineConfig({
   server: {
     proxy: {
       '/cns': {
-        target: 'http://100.127.36.37:8080',
+        target: 'http://localhost:8080',
         changeOrigin: true,
         secure: false,
         rewrite: (path) => path.replace(/^\/cns/, '')
